requirements.txt
streamlit
pillow
torch
transformers
numpy
pymilvus
